Этап 2. Основной сервис. Устранение замечаний ревью

+++ ewm-main-service/src/main/java/ru/practicum/main/dto/EventShortDto.java
@@ -39,6 +39,6 @@ public class EventShortDto {
     @Size(min = 3, max = 120)
     private String title;
 
-    private Long views;
+    private Integer views;
 }
